Methods to decompose and solve the SinglePhase Branch Flow Model are provided based on the work in [2]. These methods are most advantageous when solving the non-linear (unrelaxed) power flow equations and are only valid(?) in radial networks.

BranchFlowModel.init_inputs! — Function

init_inputs!(ps::Vector{Inputs{SinglePhase}}; init_vs::Dict = Dict())

Set the load on the upstream leaf noades equal to the sum of all the loads in the downstream inputs. It is important that the order of ps is from leaf branches to trunk branches so that the sums of loads take into account all downstream sub-trees.

if init_vs is provided, the p.v0 is set for the Input with its p.substation_bus equal to the key in init_vs

init_vs = Dict(
    "sub_bus_1" => 0.98
)

for p in ps
    if p.substation_bus in keys(init_vs)
        p.v0 = init_vs[p.substation_bus]
    end
end

init_inputs!(mg::MetaDiGraph; init_vs::Dict = Dict())

Use the :load_sum_order in mg to init_inputs! in the correct order, i.e. set the loads at the leaf - substation connections as sums of all the loads (and the voltages at substations)

BranchFlowModel.set_inputs! — Function

set_inputs!(mg::MetaDiGraph; α::Float64=0.0)

Set the shared values in each subgraph / vertex of mg:

set the current vertex's v0 to its inneighbor's voltage

set the current vertex P/Qload to the outneighbors' substation_bus loads

BranchFlowModel.split_at_busses — Function

split_at_busses(p::Inputs{SinglePhase}, at_busses::Vector{String})

Split Inputs using the at_busses

returns MetaDiGraph with vertex properties :p containing Input for the sub-graphs. For example mg[2, :p] is the Input at the second vertex of the graph created by splitting the network via the at_busses.

split_at_busses(p::Inputs{SinglePhase}, at_busses::Vector{String}, with_busses::Vector{Vector{String}})

Split up p using the at_busses as each new substation_bus and containing the corresponding with_busses. The at_busses and with_busses can be determined using splitting_busses.

NOTE: this variation of spltatbusses allows for more than two splits at the same bus; whereas the other implementation of splitatbusses only splits the network into two parts for everything above and everything below a splitting bus.

BranchFlowModel.get_diffs — Function

get_diffs(mg::MetaDiGraph)

Uses the JuMP Models stored in mg[:m] to calculate the difference between Pj, Qj, and |v| at every leaf/substation connection.

returns three Float64[]

BranchFlowModel.splitting_busses — Function

splitting_busses(p::Inputs{SinglePhase}, source::String; threshold::Int64=10)

Determine the busses to split a tree graph on by searching upward from the deepest leafs first and gathering the nearest busses until threshold is met for each subgraph.

Returns a Vector{String} for the bus names to split on and Vector{Vector{String}} for the corresponding busses within each sub-graph.

Note

It is not enough to have only the splitting busses to obey the max_busses limit because one must also know which sub branches to take from each splitting bus. In other words, we also need all the busses within each subgraph to split properly. For example, if a splitting bus has two sub branches then obeying the max_busses limit can require only including one sub branch out of the splitting bus. To know which branch to take we can use the other busses in the sub graph (which is why this method also returns the bussed in each subgraph).

BranchFlowModel.connect_subgraphs_at_busses — Function

connect_subgraphs_at_busses(p::Inputs{SinglePhase}, at_busses::Vector{String}, subgraphs::Vector{Vector})

The splitting_busses algorithm does not include over laps in subgraphs. But, we want overlaps at the splitting busses for solving the decomposed branch flow model. So here we add the overlapping splitting busses to each sub graph.

Missing docstring.

Missing docstring for split_inputs. Check Documenter's build log for details.

BranchFlowModel.build_metagraph — Function

build_metagraph(p::Inputs{SinglePhase}, source::String; max_busses::Int64=10)

return MetaDiGraph with :p property set for every vertex by splitting the Inputs via splitting_busses

[2]

Sadnan, Rabayet, and Anamika Dubey. "Distributed optimization using reduced network equivalents for radial power distribution systems." IEEE Transactions on Power Systems 36.4 (2021): 3645-3656.