BranchFlowModel.jl

BranchFlowModel builds the branch flow constraints using JuMP. The intent of this package is to allow users to build mathematical programs that include BranchFlowModel constraints. No objective is added to the JuMP model in this package and so solving any problem defined by the constraints built by BranchFlowModel.jl is a feasibility problem. Dictionaries of constraints are provided so that one can delete and/or modify the base constraints to fit their problem.

Warning

This package is under development. Contributions are welcome via fork and pull request.

Inputs

Inputs are defined using CommonOPF.Network structs.

Building a Model

Building a BranchFlowModel requires three things:

a JuMP Model,

a CommonOPF.Network, and

the type of model to be built, i.e. one of the BranchFlowModel.ModelType

BranchFlowModel.ModelType — Type

ModelType

An enum with values:

Unrelaxed

AngleRelaxation

Semidefinite

SOC

Linear

Note

JuMP exports SecondOrderCone so we abbreviate it as SOC.

To build a model see build_bfm!