Methods to decompose and solve the SinglePhase Branch Flow Model are provided based on the work in [2]. These methods are most advantageous when solving the non-linear (unrelaxed) power flow equations and are only valid(?) in radial networks.

Missing docstring.

Missing docstring for init_inputs!. Check Documenter's build log for details.

BranchFlowModel.set_inputs! — Function

set_inputs!(mg::MetaGraphsNext.MetaGraph; α::Float64=0.0)

Set the shared values in each subgraph / vertex of mg:

set the current vertex's v0 to its inneighbor's voltage

set the current vertex P/Qload to the outneighbors' substation_bus loads

BranchFlowModel.get_diffs — Function

get_diffs(mg::MetaGraphsNext.MetaGraph)

Uses the JuMP Models stored in mg.graph_data[:models] to calculate the difference between power injections/loads, and |v| at every leaf/substation connection.

returns three Float64[]

Missing docstring.

Missing docstring for split_inputs. Check Documenter's build log for details.

[2]

Sadnan, Rabayet, and Anamika Dubey. "Distributed optimization using reduced network equivalents for radial power distribution systems." IEEE Transactions on Power Systems 36.4 (2021): 3645-3656.