Methods

Some various methods used in BranchFlowModel.jl:

Warning

This list of exported methods may not be up to date and there are missing doc strings. Contributions are welcome via fork and pull request.

Missing docstring.

Missing docstring for Inputs(::String, ::String). Check Documenter's build log for details.

CommonOPF.Inputs — Method

Inputs(
    edges::Array{Tuple}, 
    linecodes::Array{String}, 
    linelengths::Array{Float64}, 
    phases::Vector{Vector},
    substation_bus::String;
    Pload, 
    Qload, 
    Sbase=1, 
    Vbase=1, 
    Zdict, 
    v0, 
    v_lolim=0.95, 
    v_uplim=1.05,
    Ntimesteps=1, 
    P_up_bound=1e4,
    Q_up_bound=1e4,
    P_lo_bound=-1e4,
    Q_lo_bound=-1e4,
    Isquared_up_bounds=Dict{String, Float64}(),
    relaxed=true
)

Lowest level Inputs constructor (the only one that returns the Inputs struct).

Note

The real and reactive loads provided are normalized using Sbase.

CommonOPF.singlephase38linesInputs — Function

singlephase38linesInputs(;
    Pload=Dict{String, AbstractArray{Real, 1}}(), 
    Qload=Dict{String, AbstractArray{Real, 1}}(), 
    T=24,
    loadnodes = ["3", "5", "36", "9", "10", "11", "12", "13", "15", "17", "18", "19", "22", "25", 
                "27", "28", "30", "31", "32", "33", "34", "35"],
    Sbase = 1e6,
    Vbase = 12.5e3,
    v0=1.0,
    v_uplim = 1.05,
    v_lolim = 0.95,
)

Convenience function for creating a single phase network with 38 lines and nodes. Taken from: Andrianesis et al. 2019 "Locational Marginal Value of Distributed Energy Resources as Non-Wires Alternatives"

NOTE that Inputs is a mutable struct (s.t. loads can be added later).

CommonOPF.dsstxt_to_sparse_array — Function

dsstxt_to_sparse_array(fp::String, first_data_row::Int = 5)

convert a SystemY.txt file from OpenDSS to a julia matrix. assumes that Y is symmetric.

Missing docstring.

Missing docstring for dss_files_to_dict. Check Documenter's build log for details.

BranchFlowModel.build_model! — Method

build_model!(m::JuMP.AbstractModel, p::Inputs{SinglePhase})

Add variables and constraints to m using the values in p. Calls the following functions:

add_variables(m, p)
constrain_power_balance(m, p)
constrain_substation_voltage(m, p)
constrain_KVL(m, p)
constrain_cone(m, p)
constrain_loads(m, p)

BranchFlowModel.build_model! — Method

build_model!(m::JuMP.AbstractModel, p::Inputs{MultiPhase})

Add variables and constraints to m using the values in p. Calls the following functions:

add_variables(m, p)
constrain_power_balance(m, p)
constrain_substation_voltage(m, p)
constrain_KVL(m, p)
constrain_loads(m, p)

BranchFlowModel.add_variables — Function

add_variables(m, p::Inputs{MultiPhase})

Create complex variables:

m[:w] are 3x3 Hermitian matrices of voltage squared (V*V^T)

m[:l] are 3x3 Hermitian matrices of current squared (I*I^T)

m[:Sj] are 3x1 matrices of net power injections (at bus j)

m[:Sij] are 3x3 Complex matrices of line flow powers (from i to j)

The positive semi-definite constraints are also defined and stored as

m[:H][t][j] where t is time step and j is the bus name.

All of the variable containers have typeof Dict{Int, Dict{String, AbstractVecOrMat}}`.

The first index is time step (integer)

The second index is bus or line (string)

and finally a matrix of complex variables

Some examples of using variables:

value.(m[:Sj][1]["671"])

value(variable_by_name(m, "real(Sj_1_671_1)"))

value.(m[:w][1]["671"])

value.(m[:Sj][1][p.substation_bus]) 

for b in keys(p.Pload)
    println(b, "  ", value.(m[:Sj][1][b]))
end

fix(variable_by_name(m, "real(Sj_1_645_3)"), 0.0, force=true)

BranchFlowModel.constrain_power_balance — Function

function constrain_power_balance(m, p::Inputs)

Pij in - losses == sum of line flows out + net injection NOTE: using sum over Pij for future expansion to mesh grids i -> j -> k

function constrain_power_balance(m, p::Inputs{MultiPhase})

Sij in - losses == sum of line flows out + net injection NOTE: using sum over Pij for future expansion to mesh grids i -> j -> k

All of the power balance constraints are stored in m[:loadbalcons] with the bus name (string) as the first index. For example m[:loadbalcons]["busname"] will give the constrain container from JuMP for all time steps.

Missing docstring.

Missing docstring for constrain_substation_voltage. Check Documenter's build log for details.

BranchFlowModel.constrain_KVL — Function

constrain_KVL(m, p::Inputs{MultiPhase})

Add the voltage drop definintions between busses.

BranchFlowModel.constrain_loads — Function

constrain_loads(m, p::Inputs)

set net injections Pj/Qj to negative of Inputs.Pload/Qload, which are normalized by Sbase when creating Inputs

keys of P/Qload must match Inputs.busses. Any missing keys have load set to zero.

Inputs.substation_bus is unconstrained, slack bus

constrain_loads(m, p::Inputs{MultiPhase})

set loads to negative of Inputs.Pload and Inputs.Qload, which are normalized by Sbase when creating Inputs.

keys of Pload and Qload must match Inputs.busses. Any missing keys have load set to zero.

Inputs.substation_bus is unconstrained, slack bus

Each of the power injection constraints are stored in the model under m[:injectioncons]. To acces the constraints you can:

m[:injectioncons]["busname"][t,phs]

where t is the integer time step and phs is the integer phase.

Missing docstring.

Missing docstring for constrain_bounds. Check Documenter's build log for details.

BranchFlowModel.check_rank_one — Function

check_rank_one(m::JuMP.AbstractModel, p::Inputs{BranchFlowModel.MultiPhase}, tol=1e-3)

Check the rank of the m[:H] matrices from the PSD cone constraints. Warnings express any values with rank greater than one.

Missing docstring.

Missing docstring for get_bus_values. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_edge_values. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_ijlinecode. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_ijlinelength. Check Documenter's build log for details.

CommonOPF.get_ij_idx — Function

get_ij_idx(i::AbstractString, j::AbstractString, p::Inputs)

get the index for edge i->j in the Inputsedge vectors:edges,linecodes,phases,edge_keys, andlinelengths`

CommonOPF.i_to_j — Function

function i_to_j(j::AbstractString, p::Inputs)

find all busses upstream of bus j

Note

In a radial network this function should return an Array with length of 1.

CommonOPF.j_to_k — Function

function j_to_k(j::AbstractString, p::Inputs)

find all busses downstream of bus j

BranchFlowModel.rij — Function

rij(i::AbstractString, j::AbstractString, p::Inputs{SinglePhase})

The per-unit resistance of line i->j

BranchFlowModel.xij — Function

xij(i::AbstractString, j::AbstractString, p::Inputs{SinglePhase})

The per-unit reacttance of line i->j

Missing docstring.

Missing docstring for zij. Check Documenter's build log for details.

BranchFlowModel.check_soc_inequalities — Function

check_soc_inequalities(m::JuMP.AbstractModel, p::Inputs)

create and return a vector of the gaps in the second order cone constraints

BranchFlowModel.check_connected_graph — Function

check_connected_graph(p::Inputs{SinglePhase})

return true if only one connected graph; false otherwise

this is a good check to do before attempting to solve a model because if there is more than one sub-graph then it is likely the model will be infeasible

BranchFlowModel.get_load_bal_shadow_prices — Function

get_load_bal_shadow_prices(m::JuMP.AbstractModel, p::Inputs)

create and return a dict indexed by bus and time for shadow prices (just real prices for now)

Missing docstring.

Missing docstring for voltage_values_by_time_bus. Check Documenter's build log for details.

Missing docstring.

Missing docstring for current_values_by_time_edge. Check Documenter's build log for details.

Missing docstring.

Missing docstring for line_flow_values_by_time_edge. Check Documenter's build log for details.

BranchFlowModel.reduce_tree! — Function

reduce_tree!(p::Inputs{SinglePhase})

combine any line sets with intermediate busses that have indegree == outdegree == 1 and is not a load bus into a single line

Missing docstring.

Missing docstring for trim_tree!. Check Documenter's build log for details.

CommonOPF.make_graph — Function

make_graph(busses::AbstractVector{String}, edges::AbstractVector)

return SimpleDiGraph, Dict, Dict with the dicts for bus => int and int => bus (because Graphs.jl only works with integer nodes)

julia> g["13", :bus]
10

julia> get_prop(g, :bus_int_map)["13"]
10

julia> g[13, :bus]
"24"

julia> get_prop(g, :int_bus_map)[13]
"24"

BranchFlowModel.leaf_busses — Function

leaf_busses(p::Inputs)

returns Vector{String} containing all of the leaf busses in p.busses

BranchFlowModel.init_inputs! — Function

init_inputs!(ps::Vector{Inputs{SinglePhase}}; init_vs::Dict = Dict())

Set the load on the upstream leaf noades equal to the sum of all the loads in the downstream inputs. It is important that the order of ps is from leaf branches to trunk branches so that the sums of loads take into account all downstream sub-trees.

if init_vs is provided, the p.v0 is set for the Input with its p.substation_bus equal to the key in init_vs

init_vs = Dict(
    "sub_bus_1" => 0.98
)

for p in ps
    if p.substation_bus in keys(init_vs)
        p.v0 = init_vs[p.substation_bus]
    end
end

init_inputs!(mg::MetaDiGraph; init_vs::Dict = Dict())

Use the :load_sum_order in mg to init_inputs! in the correct order, i.e. set the loads at the leaf - substation connections as sums of all the loads (and the voltages at substations)

BranchFlowModel.set_inputs! — Function

set_inputs!(mg::MetaDiGraph; α::Float64=0.0)

Set the shared values in each subgraph / vertex of mg:

set the current vertex's v0 to its inneighbor's voltage

set the current vertex P/Qload to the outneighbors' substation_bus loads

BranchFlowModel.split_at_busses — Function

split_at_busses(p::Inputs{SinglePhase}, at_busses::Vector{String})

Split Inputs using the at_busses

returns MetaDiGraph with vertex properties :p containing Input for the sub-graphs. For example mg[2, :p] is the Input at the second vertex of the graph created by splitting the network via the at_busses.

split_at_busses(p::Inputs{SinglePhase}, at_busses::Vector{String}, with_busses::Vector{Vector{String}})

Split up p using the at_busses as each new substation_bus and containing the corresponding with_busses. The at_busses and with_busses can be determined using splitting_busses.

NOTE: this variation of spltatbusses allows for more than two splits at the same bus; whereas the other implementation of splitatbusses only splits the network into two parts for everything above and everything below a splitting bus.

BranchFlowModel.get_diffs — Function

get_diffs(mg::MetaDiGraph)

Uses the JuMP Models stored in mg[:m] to calculate the difference between Pj, Qj, and |v| at every leaf/substation connection.

returns three Float64[]

Missing docstring.

Missing docstring for all_outneighbors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for all_inneighbors. Check Documenter's build log for details.

CommonOPF.busses_from_deepest_to_source — Function

busses_from_deepest_to_source(g::MetaDiGraph, source::String)

return the busses and their integer depths in order from deepest from shallowest

Missing docstring.

Missing docstring for vertices_from_deepest_to_source. Check Documenter's build log for details.

BranchFlowModel.splitting_busses — Function

splitting_busses(p::Inputs{SinglePhase}, source::String; threshold::Int64=10)

Determine the busses to split a tree graph on by searching upward from the deepest leafs first and gathering the nearest busses until threshold is met for each subgraph.

Returns a Vector{String} for the bus names and Vector{Vector{String}} for the corresponding busses within each sub-graph.

Note

It is not enough to have only the splitting busses to obey the max_busses limit because one must also know which sub branches to take from each splitting bus. In other words, we also need all the busses within each subgraph to split properly. For example, if a splitting bus has two sub branches then obeying the max_busses limit can require only including one sub branch out of the splitting bus. To know which branch to take we can use the other busses in the sub graph.

BranchFlowModel.connect_subgraphs_at_busses — Function

connect_subgraphs_at_busses(p::Inputs{SinglePhase}, at_busses::Vector{String}, subgraphs::Vector{Vector})

The splitting_busses algorithm does not include over laps in subgraphs. But, we want overlaps at the splitting busses for solving the decomposed branch flow model. So here we add the overlapping splitting busses to each sub graph.

BranchFlowModel.split_inputs — Function

split_inputs(p::Inputs{SinglePhase}, bus::String, g::SimpleDiGraph)

Split inputs into one graph for everything above bus and one graph for everything below bus.

split_inputs(p::Inputs{SinglePhase}, bus::String, out_buses::Vector{String})

Split p into p_above and p_below where p_below has only out_buses and p_above has union( [bus], setdiff(p.busses, out_buses) ).

Note that out_buses must contain bus

BranchFlowModel.build_metagraph — Function

build_metagraph(p::Inputs{SinglePhase}, source::String; max_busses::Int64=10)

return MetaDiGraph with :p property set for every vertex by splitting the Inputs via splitting_busses

BranchFlowModel.solve_metagraph! — Function

solve_metagraph!(mg::MetaDiGraph, builder::Function, tol::T; α::T=0.5, verbose=false) where T <: Real

Given a MetaDiGraph and a JuMP Model builder method iteratively solve the models until the tol is met for the differences provided by BranchFlowModel.get_diffs.

The builder must accept only one argument of type BranchFlowModel.AbstractInputs that returns a JuMP.AbstractModel. Each model returned from the builder is stored as an :m property in each vertex of mg.

Note

tol is compared to the maximum absolute value of all the p, q, and v differences.

solve_metagraph!(mg::MetaDiGraph, builder::Function, tols::Vector{T}; α::T=0.5, verbose=false) where T <: Real

Given a MetaDiGraph and a JuMP Model builder method iteratively solve the models until the tols are met for the differences provided by BranchFlowModel.get_diffs.

The builder must accept only one argument of type BranchFlowModel.AbstractInputs that returns a JuMP.AbstractModel. Each model returned from the builder is stored as an :m property in each vertex of mg.

Note

The tols should have a length of three. The first value is compared to the maximum absolute difference in Pj, the second for Qj, and the third for |v|. All differences are calculated at the leaf/substation connections.

solve_metagraph!(mg::MetaDiGraph, builder::Dict{Int64, Function}, tols::Vector{T}; α::T=0.5, verbose=false) where T <: Real

Given a MetaDiGraph and a JuMP Model builder method iteratively solve the models until the tols are met for the differences provided by BranchFlowModel.get_diffs. The builder dict is used to build each model for the corresponding vertex key.

Each function in the builder dict must accept only one argument of type BranchFlowModel.AbstractInputs that returns a JuMP.AbstractModel. Each model returned from the builder function is stored as an :m property in each vertex of mg.

Note

The tols should have a length of three. The first value is compared to the maximum absolute difference in Pj, the second for Qj, and the third for |v|. All differences are calculated at the leaf/substation connections.

Missing docstring.

Missing docstring for metagraph_voltages. Check Documenter's build log for details.

BranchFlowModel.check_unique_solution_conditions — Function

check_unique_solution_conditions(p::Inputs)

report the maximum per-unit immpedance and load values. See Chiang and Baran 2013: A load flow solution with feasible voltage magnitude always exists and is unique when

V0 ≈ 1

loss values < 1

rpu, xpu << 1

Missing docstring.

Missing docstring for check_statuses. Check Documenter's build log for details.

BranchFlowModel.reg_busses — Function

reg_busses(p::Inputs)

All of the regulated busses, i.e. the second bus in the regulated edges

Missing docstring.

Missing docstring for turn_ratio. Check Documenter's build log for details.

Missing docstring.

Missing docstring for vreg. Check Documenter's build log for details.

Missing docstring.

Missing docstring for has_vreg. Check Documenter's build log for details.

BranchFlowModel.combine_parallel_lines! — Function

combine_parallel_lines!(p::Inputs)

Combine any parallel single phase lines without loads on intermediate busses into one multiphase line

TODO this can probably be used for LinDistFlow too but then have to move rij and other utils.jl stuff to CommonOPF.jl (and distinguish between multiphase LinDistFlow and BranchFlowModel Inputs)

BranchFlowModel.remove_bus! — Function

remove_bus!(j::String, p::Inputs{SinglePhase})

Remove bus j in the line i->j->k from the model by making an equivalent line from busses i->k

remove_bus!(j::String, p::Inputs{MultiPhase})

Remove bus j in the line i->j->k from the model by making an equivalent line from busses i->k

Missing docstring.

Missing docstring for paths_between. Check Documenter's build log for details.

BranchFlowModel.Results — Method

Results(m::AbstractModel, p::Inputs{SinglePhase})

return a Results struct with fieldnames:

voltage_magnitudes
real_power_injections
reactive_power_injections
current_magnitudes
real_sending_end_powers
reactive_sending_end_powers