Methods

Some various methods used in BranchFlowModel.jl:

Warning

This list of exported methods may not be up to date and there are missing doc strings. Contributions are welcome via fork and pull request.

Missing docstring.

Missing docstring for Inputs(::String, ::String). Check Documenter's build log for details.

CommonOPF.Inputs — Method

Inputs(
    edges::Array{Tuple}, 
    linecodes::Array{String}, 
    linelengths::Array{Float64}, 
    phases::Vector{Vector},
    substation_bus::String;
    Pload, 
    Qload, 
    Sbase=1, 
    Vbase=1, 
    Zdict, 
    v0, 
    v_lolim=0.95, 
    v_uplim=1.05,
    Ntimesteps=1, 
    P_up_bound=1e4,
    Q_up_bound=1e4,
    P_lo_bound=-1e4,
    Q_lo_bound=-1e4,
    Isquared_up_bounds=Dict{String, Float64}(),
    relaxed=true
)

Lowest level Inputs constructor (the only one that returns the Inputs struct).

Note

The real and reactive loads provided are normalized using Sbase.

CommonOPF.singlephase38linesInputs — Function

singlephase38linesInputs(;
    Pload=Dict{String, AbstractArray{Real, 1}}(), 
    Qload=Dict{String, AbstractArray{Real, 1}}(), 
    T=24,
    loadnodes = ["3", "5", "36", "9", "10", "11", "12", "13", "15", "17", "18", "19", "22", "25", 
                "27", "28", "30", "31", "32", "33", "34", "35"],
    Sbase = 1e6,
    Vbase = 12.5e3,
    v0=1.0,
    v_uplim = 1.05,
    v_lolim = 0.95,
)

Convenience function for creating a single phase network with 38 lines and nodes. Taken from: Andrianesis et al. 2019 "Locational Marginal Value of Distributed Energy Resources as Non-Wires Alternatives"

NOTE that Inputs is a mutable struct (s.t. loads can be added later).

CommonOPF.dsstxt_to_sparse_array — Function

dsstxt_to_sparse_array(fp::String, first_data_row::Int = 5)

convert a SystemY.txt file from OpenDSS to a julia matrix. assumes that Y is symmetric.

Missing docstring.

Missing docstring for dss_files_to_dict. Check Documenter's build log for details.

Missing docstring.

Missing docstring for build_model!(m::JuMP.AbstractModel, p::Inputs{BranchFlowModel.SinglePhase}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for build_model!(m::JuMP.AbstractModel, p::Inputs{BranchFlowModel.MultiPhase}). Check Documenter's build log for details.

BranchFlowModel.add_variables — Function

add_variables(m, net::Network{MultiPhase}; PSD::Bool=true)

Create complex variables:

m[:w] are 3x3 Hermitian matrices of voltage squared (V*V^T)

m[:l] are 3x3 Hermitian matrices of current squared (I*I^T)

m[:Sj] are 3x1 matrices of net power injections (at bus j)

m[:Sij] are 3x3 Complex matrices of line flow powers (from i to j)

If PSD is true then the positive semi-definite constraints are also defined and stored as

m[:H][t][j] where t is time step and j is the bus name.

All of the variable containers have typeof Dict{Int, Dict{String, AbstractVecOrMat}}`.

The first index is time step (integer)

The second index is bus or line (string)

and finally a matrix or vector of complex variables

Some examples of using variables:

value.(m[:Sj][1]["671"])

value(variable_by_name(m, "real(Sj_1_671_1)"))

value.(m[:w][1]["671"])

value.(m[:Sj][1][net.substation_bus]) 

for b in real_load_busses(net)
    println(b, "  ", value.(m[:Sj][1][b]))
end

fix(variable_by_name(m, "real(Sj_1_645_3)"), 0.0, force=true)

BranchFlowModel.constrain_power_balance — Function

function constrain_power_balance(m, net::Network)

Define the m[:loadbalcons][bus] ∀ bus ∈ busses(net) as a Dict of constraints. The keys are "p" and "q" for real and reactive power balance respectively. The values are the JuMP constraints.

∑ Pij in - losses + net injection - ∑ Pjk out = 0

The net injection are user defined loads. If one wishes to make the net injection a decision variable then delete the constraint and redefine the constraint with your decision variable.

NOTE: using sum over Pij for future expansion to mesh grids and the convention: i -> j -> k

function constrain_power_balance(m, net::Network{MultiPhase})

Sij in - losses == sum of line flows out + net injection NOTE: using sum over Pij for future expansion to mesh grids i -> j -> k

All of the power balance constraints are stored in m[:loadbalcons] with the bus name (string) as the first index. For example m[:loadbalcons]["busname"] will give the constrain container from JuMP for all time steps.

Missing docstring.

Missing docstring for constrain_substation_voltage. Check Documenter's build log for details.

BranchFlowModel.constrain_KVL — Function

constrain_KVL(m, net::Network{MultiPhase})

Add the voltage drop definitions between busses.

$w_j = w_i - S_{ij} Z^{\star} - Z S_{ij}^{\star} + Z L_{ij} Z^{\star}$

Missing docstring.

Missing docstring for constrain_bounds. Check Documenter's build log for details.

Missing docstring.

Missing docstring for check_rank_one. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_bus_values. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_edge_values. Check Documenter's build log for details.

Missing docstring.

Missing docstring for check_soc_inequalities. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_load_bal_shadow_prices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for current_values_by_time_edge. Check Documenter's build log for details.

Missing docstring.

Missing docstring for line_flow_values_by_time_edge. Check Documenter's build log for details.

CommonOPF.reduce_tree! — Function

reduce_tree!(p::Inputs{SinglePhase})

combine any line sets with intermediate busses that have indegree == outdegree == 1 and is not a load bus into a single line

See remove_bus! for how the two lines are combined.

reduce_tree!(p::Inputs{MultiPhase})

combine any line sets with intermediate busses that satisfy 1. indegree == outdegree == 1 2. is not a load bus and 3. has same phases in as out into a single line.

See remove_bus! for how the two lines are combined.

CommonOPF.trim_tree! — Function

trim_tree!(p::Inputs)

Trim any branches that do not contain load busses.

CommonOPF.make_graph — Function

make_graph(edges::AbstractVector)

return SimpleDiGraph by inferring busses from edges with the dicts for bus => int and int => bus (because Graphs.jl only works with integer nodes)

julia> g["13", :bus]
10

julia> g[13, :bus]
"24"

julia> get_prop(g, :int_bus_map)[13]
"24"

make_graph(busses::AbstractVector{String}, edges::AbstractVector)

return SimpleDiGraph with the dicts for bus => int and int => bus (because Graphs.jl only works with integer nodes)

julia> g["13", :bus]
10

julia> g[13, :bus]
"24"

julia> get_prop(g, :int_bus_map)[13]
"24"

CommonOPF.leaf_busses — Function

leaf_busses(p::Inputs)

returns Vector{String} containing all of the leaf busses in p.busses

Missing docstring.

Missing docstring for set_inputs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_diffs. Check Documenter's build log for details.

BranchFlowModel.solve_metagraph! — Function

solve_metagraph!(mg::MetaGraphsNext.MetaGraph, builder::Function, tol::T; α::T=0.5, verbose=false) where T <: Real

Given a MetaGraphsNext.MetaGraph and a JuMP Model builder method iteratively solve the models until the tol is met for the differences provided by BranchFlowModel.get_diffs.

The builder must accept only one argument of type CommonOPF.AbstractNetwork that returns a JuMP.AbstractModel. Each model returned from the builder is stored as an :m property in each vertex of mg.

Note

tol is compared to the maximum absolute value of all the p, q, and v differences.

solve_metagraph!(mg::MetaGraphsNext.MetaGraph, builder::Function, tols::Vector{T}; α::T=0.5, verbose=false) where T <: Real

Given a MetaGraphsNext.MetaGraph and a JuMP Model builder method iteratively solve the models until the tols are met for the differences provided by BranchFlowModel.get_diffs.

The builder must accept only one argument of type CommonOPF.AbstractNetwork that returns a JuMP.AbstractModel. Each model returned from the builder is stored as an :m property in each vertex of mg.

Note

The tols should have a length of three. The first value is compared to the maximum absolute difference in real power, the second for reactive power, and the third for |v|. All differences are calculated at the leaf/substation connections.

solve_metagraph!(mg::MetaGraphsNext.MetaGraph, builder::Dict{Int64, Function}, tols::Vector{T}; α::T=0.5, verbose=false) where T <: Real

Given a MetaGraphsNext.MetaGraph and a JuMP Model builder method iteratively solve the models until the tols are met for the differences provided by BranchFlowModel.get_diffs. The builder dict is used to build each model for the corresponding vertex key.

Each function in the builder dict must accept only one argument of type CommonOPF.AbstractNetwork that returns a JuMP.AbstractModel. Each model returned from the builder function is stored as an :m property in each vertex of mg.

Note

The tols should have a length of three. The first value is compared to the maximum absolute difference in real power, the second for reactive power, and the third for |v|. All differences are calculated at the leaf/substation connections.

Missing docstring.

Missing docstring for metagraph_voltages. Check Documenter's build log for details.

Missing docstring.

Missing docstring for check_unique_solution_conditions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for check_statuses. Check Documenter's build log for details.

CommonOPF.reg_busses — Function

reg_busses(p::Inputs)

All of the regulated busses, i.e. the second bus in the regulated edges

Missing docstring.

Missing docstring for turn_ratio. Check Documenter's build log for details.

Missing docstring.

Missing docstring for vreg. Check Documenter's build log for details.

Missing docstring.

Missing docstring for has_vreg. Check Documenter's build log for details.

CommonOPF.remove_bus! — Function

remove_bus!(j::String, p::Inputs{SinglePhase})

Remove bus j in the line i->j->k from the model by making an equivalent line from busses i->k

remove_bus!(j::String, p::Inputs{MultiPhase})

Remove bus j in the line i->j->k from the model by making an equivalent line from busses i->k

Missing docstring.

Missing docstring for paths_between. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Results(m::JuMP.AbstractModel, p::Inputs{SinglePhase}; digits=8). Check Documenter's build log for details.